+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ## [Unreleased]
 ### Added
+- Prevent JetBrains AI Assistance using a .noai file
+
+### Changed
+- Plugin now restricted to IntelliJ IDEA only (removed support for other JetBrains IDEs)
+
+
+## [0.1.0] - 2025-09-18
+### Added
 - Disable red/yellow light bulb icons that appear next to code issues
 - Disables quick previews of potential fixes
 - Detect the presence of the GitHub Copilot plugin, Gemini Code Assist and DeepSeek AI developer